requirements.txt
adafruit-io
autopep8
django-cors-headers
django-environ
django-filter
djangorestframework-simplejwt
flake8
gunicorn
markdown
pip-chill
pysqlite3
whitenoise
